+++ KisakCOD/src/game/bullet.h
@@ -18,7 +18,7 @@ char __cdecl BulletTrace(
     const centity_s *attacker,
     BulletTraceResults *br,
     unsigned int lastSurfaceType);
-double __cdecl G_GoodRandomFloat(int *idum);
+float __cdecl G_GoodRandomFloat(int *idum);
 void __cdecl Bullet_Endpos(int randSeed, float spread, float *end, float *dir, const weaponParms *wp, float maxRange);
 void __cdecl Bullet_RandomDir(int time, float *x, float *y);
 void __cdecl Bullet_Fire(
